+++ internal/host/memory/memory.go
