+++ internal/host/memory/type.go
