+++ pkg/host/host_type.go
@@ -1,0 +1,13 @@
+package host
+
+import (
+	"github.com/CharVstack/CharV-lib/internal/host/cpu"
+	"github.com/CharVstack/CharV-lib/internal/host/memory"
+	"github.com/CharVstack/CharV-lib/internal/host/storage"
+)
+
+type Host struct {
+	Cpu          cpu.Cpu
+	Memory       memory.Memory
+	StoragePools []*storage.PoolInfo
+}
