+++ pkg/qemu/install.go
