+++ pkg/qemu/parse.go
