+++ pkg/qemu/parse_test.go
