+++ pkg/qemu/type.go
